app/order/views.py
from django.shortcuts import render
from rest_framework import generics

from .models import OrderItem, Order
from .serializers import OrderItemSerializer, OrderSerializer


class OrderItemListCreateAPIView(generics.ListCreateAPIView):
    queryset = OrderItem.objects.all()
    serializer_class = OrderItemSerializer


class OrderListCreateAPIView(generics.ListCreateAPIView):
    queryset = Order.objects.all()
    serializer_class = OrderSerializer
