app/requirements.txt
django==3.2
djangorestframework
Pillow
drf-yasg
psycopg2-binary
gunicorn
